Skip Windows builds
Rationale: Windows was never considered a supported platform of folly.
The maintainance of the feedstock for Windows is costly while the
use of it is questionable.

folly is used by a few projects on conda-forge, some of which
aren't distributed for Windows anyway.

See: https://github.com/facebook/folly?tab=readme-ov-file#build-notes
See: https://github.com/facebook/folly/issues/962

Signed-off-by: Julien Jerphanion <git@jjerphan.xyz>